Implementação do tratamento de erros para salvar ou excluir uma ordem de serviço.

+++ src/main/java/com/danrocha/dialogframework/services/OrdemServicoService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import java.io.Serial;
 import java.io.Serializable;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class OrdemServicoService implements Serializable {
@@ -17,7 +18,10 @@ public class OrdemServicoService implements Serializable {
     @Autowired
     private transient OrdemServicoRepository ordemRepo;
 
-    public void salvarOuAtualizar(OrdemServico ordem) {
+    public void salvarOuAtualizar(OrdemServico ordem) throws NegocioException {
+        if (Objects.isNull(ordem)) {
+            throw new NegocioException("Não foi possível salvar ou atualizar a Ordem de Serviço");
+        }
         this.ordemRepo.saveAndFlush(ordem);
     }
 
@@ -26,7 +30,10 @@ public class OrdemServicoService implements Serializable {
         return this.ordemRepo.findAll();
     }
 
-    public void excluir(OrdemServico ordem) {
+    public void excluir(OrdemServico ordem) throws NegocioException {
+        if (Objects.isNull(ordem)) {
+            throw new NegocioException("Não foi possível excluir a Ordem de Serviço");
+        }
         this.ordemRepo.delete(ordem);
     }
 }
